Commit correções nome de classes e remoção feiranteController e OrganizadorController
