--- src/main/java/br/com/feirasverdes/backend/controller/OrganizadorController.java
@@ -1,42 +1,0 @@
-package br.com.feirasverdes.backend.controller;
-
-import java.util.List;
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import br.com.feirasverdes.backend.dao.UsuarioDao;
-import br.com.feirasverdes.backend.entidade.Usuario;
-
-@RestController
-@CrossOrigin
-@RequestMapping(value = "/organizador")
-public class OrganizadorController {
-
-	@Autowired
-	private UsuarioDao dao;
-
-	@GET
-	@Produces(MediaType.APPLICATION_JSON)
-	public Response listarTodos() {
-		return Response.ok(dao.findAll()).build();
-	}
-
-	@GET
-	@Path("{nome}")
-	@Produces(MediaType.APPLICATION_JSON)
-	public Response pesquisarPorNome(@PathParam("nome") String nome) {
-		List<Usuario> usuarios = dao.pesquisarPorNome(nome);
-		return Response.ok(usuarios).build();
-	}
-
-}
